+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 * New option to truncate files to full size (--trunctosize) in creation/write phase.
 * New option to preallocate file disk space (--preallocfile) in creation/write phase.
 * File deletion (-F) can be used when benchmark path is a file.
+* New directory sharing option (--dirsharing) to make all threads work in the same dirs instead of individual dirs for each thread.
+
+### General Changes
+* Renamed option "--notshared" to "--nosvcshare" for clear separation from new "--dirsharing" option.
+* Filenames in dir mode now include the thread index (rank) to avoid conflicts in case of dir sharing.
 
 ## v1.5.3 (Nov 15, 2020)